# swagger.yaml
...
swagger: "2.0"
host: "localhost:8000"
basePath: "/api/v0.3"
info:
  title: User-Organization API
  description: An API which supports management of customers and organizations within the LEXIS portal; this version is intended for use in LEXIS Portal R1
  version: "0.3.2021-12-15"
  contact:
    email: diego@cyclops-labs.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-KEY
  APIKeyParam:
    type: apiKey
    in: query
    name: api_key
  Keycloak:
    type: oauth2
    flow:  accessCode
    authorizationUrl: 'http://localhost:8080/auth/realms/Dev/protocol/openid-connect/auth'
    tokenUrl: 'http://localhost:8080/auth/realms/Dev/protocol/openid-connect/token'
    scopes:
      lex_adm: $Deity scope, can do absoulutely anything
      lex_sup: $Deity scope with restrictions.
      org_mgr: Organization Manager
      prj_mgr: Project/Workflows Manager
      dat_mgr: Datasets Manager
      end_usr: Plain user
schemes:
  - http
  - https
security:
  - Keycloak: [lex_adm, lex_sup, org_mgr, prj_mgr, dat_mgr, end_usr]
paths:
...
  /project:
    get:
      tags:
        - projectManagement
      produces:
        - application/json
      summary: list projects
      operationId: listProjects
      description: lists all projects
      responses:
        '200':
          description: list of projectss
          schema:
            type: array
            items:
              $ref: "#/definitions/Project"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: query
          name: scope
          description: project scope switch
          type: string
          enum:
          - ALL
          - OWN
          - ALLOWED
          default: ALL
    post:
      tags:
        - projectManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: create Project
      operationId: createProject
      description: Creates a new project
      responses:
        '201':
          description: project created
          schema:
            $ref: "#/definitions/CreatedResponse"
        '409':
          description: an existing item already exists
          schema:
            $ref: "#/definitions/ConflictResponse"
        '422':
          description: Shortname duplicated
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: body
          name: project
          description: Project to be added
          schema:
            $ref: '#/definitions/Project'
  /project/{id}:
    get:
      tags:
        - projectManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Get specific project
      operationId: getProject
      description: get project with given id
      responses:
        '200':
          description: project returned
          schema:
            $ref: "#/definitions/Project"
        '404':
          description: project with not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of project to be obtained
    put:
      tags:
        - projectManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Update specific project
      operationId: updateProject
      description: Update project with given id
      responses:
        '200':
          description: updated project
          schema:
            $ref: "#/definitions/Project"
        '404':
          description: project with not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '422':
          description: Shortname duplicated
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of project to be obtained
        - in: body
          name: project
          required: true
          description: updated project data to be added
          schema:
            $ref: "#/definitions/Project"
    delete:
      tags:
        - projectManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Delete specific project
      operationId: deleteProject
      description: Delete project with given id
      responses:
        '200':
          description: deleted project
          schema:
            $ref: "#/definitions/DeletedResponse"
        '404':
          description: project with not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of project to be obtained
  /project/{id}/user/{userID}:
    put:
      tags:
        - projectManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Add user to project
      operationId: addUserToProject
      description: Add user to project
      responses:
        '200':
          description: user updated
          schema:
            $ref: "#/definitions/OKResponse"
        '404':
          description: project or user not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of project to be obtained
        - in: path
          name: userID
          type: string
          format: uuid
          required: true
          description: Id of user to be added
    delete:
      tags:
        - projectManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Delete user from project
      operationId: deleteUserFromProject
      description: Delete user from project
      responses:
        '200':
          description: deleted user from project
          schema:
            $ref: "#/definitions/DeletedResponse"
        '404':
          description: project or user not found
          schema:
            $ref: "#/definitions/MissingResponse"
        '401':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Authorization error
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of project to be obtained
        - in: path
          name: userID
          type: string
          format: uuid
          required: true
          description: Id of user to be removed
...
definitions:
...
  StringArray:
    x-go-type:
      import:
        package: "github.com/lib/pq"
      type: StringArray
...
  Project:
    type: object
    properties:
      ProjectID:
        type: string
        format: uuid
        x-go-custom-tag: gorm:"column:projectid;type:uuid;primary_key;unique;default:md5(random()::text || clock_timestamp()::text)::uuid"
      ProjectName:
        type: string
        x-go-custom-tag: gorm:"column:projectname"
      ProjectShortName:
        type: string
        x-go-custom-tag: gorm:"column:projectshortname;unique"
      ProjectDescription:
        type: string
        x-go-custom-tag: gorm:"column:projectdescription"
      ProjectCreationTime:
        type: string
        format: date-time
        x-go-custom-tag: gorm:"column:projectcreationtime;type:timestamptz;default:now()"
      ProjectCreatedBy:
        type: string
        format: uuid
        x-go-custom-tag: gorm:"column:projectcreatedby;type:uuid"
      LinkedOrganization:
        type: string
        format: uuid
        x-go-custom-tag: gorm:"column:linkedorganization;type:uuid"
      AllowedOrganizations:
        $ref: '#/definitions/StringArray'
        x-go-custom-tag: gorm:"column:allowedorganizations;type:text[]"
      ProjectStatus:
        type: string
        enum:
        - PENDING
        - ACTIVE
        - DISABLED
        - TERMINATED
        x-go-custom-tag: gorm:"column:projectstatus"
      ProjectContactPerson:
        type: string
        format: uuid
        x-go-custom-tag: gorm:"column:projectcontactperson;type:uuid"
      ProjectStartDate:
        type: string
        format: date-time
        x-go-custom-tag: gorm:"column:projectstartdate;type:timestamptz"
      ProjectTerminationDate:
        type: string
        format: date-time
        x-go-custom-tag: gorm:"column:projectterminationdate;type:timestamptz"
      ProjectMaxPrice:
        type: number
        format: double
        default: 0.0
        x-nullable: true
        x-go-custom-tag: gorm:"column:projectmaxprice;type:float8;default:0.0"
      NormCoreHours:
        type: integer
        default: 0
        x-nullable: true
        x-go-custom-tag: gorm:"column:normcorehours;default:0"
      ProjectContactEmail:
        type: string
        format: email
        x-go-custom-tag: gorm:"column:projectcontactemail"
      ProjectDomain:
        type: string
        x-go-custom-tag: gorm:"column:projectdomain"
...
  CreatedResponse:
    type: object
    properties:
      ID:
        type: string
      Link:
        type: string
  ConflictResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
  DeletedResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
  OKResponse:
    type: object
    properties:
      Message:
        type: string
  ErrorResponse:
    type: object
    properties:
      Message:
        type: string
  MissingResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
...